# openapi.yml
openapi: 3.0.3
info:
  title: HMS Backend API
  version: 1.0.0
  description: |
    OpenAPI specification based solely on the application's actual Express routes and validation rules.
  contact:
    name: HMS Admin
    email: mrfortrancer@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://hms.fortrancer.in
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload or missing required information.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required or token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller does not have the required permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong.
        statusCode:
          type: integer
          example: 500
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
        details:
          type: object
          additionalProperties: true
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.
...
    AppointmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          example: 23
        page:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
...
    PatientUpdateRequest:
      type: object
      required:
        - patientId
      properties:
        patientId:
          type: string
          example: PAT-1001
        name:
          type: string
          example: Rahul Sharma
        gender:
          type: string
          example: Male
        dob:
          type: string
          format: date
          example: 1992-05-14
        address:
          type: string
          example: 12, MG Road, Bengaluru
        bloodGroup:
          type: string
          example: O+
        allergies:
          type: string
          example: Penicillin
        emergencyContact:
          type: string
          example: +91-9998887776
...
    UserProfileResponse:
      type: object
      properties:
        message:
          type: string
          example: Sucessfully obtained user information
        name:
          type: string
          example: Dr. Asha Menon
        email:
          type: string
          format: email
          example: asha.menon@hms.com
        status:
          type: string
          example: Active
        role:
          type: string
          example: Doctor
        employeeId:
          type: string
          example: EMP-2001
        isVerified:
          type: boolean
          example: true
        firstLogin:
          type: boolean
          example: false
        department:
          type: string
          example: OPD
        designation:
          type: string
          example: Senior Consultant
        joiningDate:
          type: string
          format: date
          example: 2025-01-15
        medicalRegistrationNo:
          type: string
          example: MED-1001
        specialization:
          type: string
          example: Cardiology
        qualification:
          type: string
          example: MBBS, MD
        consultationFee:
          type: number
          example: 1000
        availabilitySlots:
          type: array
          items:
            type: string
          example: ["09:00", "11:00"]
    PatientProfileResponse:
      type: object
      properties:
        message:
          type: string
          example: Sucessfully obtained user information
        email:
          type: string
          format: email
          example: rahul.sharma@example.com
        status:
          type: string
          example: Active
        role:
          type: string
          example: Patient
        isVerified:
          type: boolean
          example: true
        uhid:
          type: string
          example: PAT-1001
        name:
          type: string
          example: Rahul Sharma
        gender:
          type: string
          example: Male
        dob:
          type: string
          format: date
          example: 1992-05-14
        allergies:
          type: string
          example: Peanuts
        bloodGroup:
          type: string
          example: O+
        address:
          type: string
          example: 12, MG Road, Bengaluru
        phone:
          type: string
          example: +91-9876543210
        emergencyContact:
          type: string
          example: +91-9998887776
...
paths:
...
  /user/getUserProfile:
    get:
      tags: [User]
      summary: Get user profile by email
      description: Returns employee profile data for the supplied email.
      operationId: getUserProfile
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getPatients:
    get:
      tags: [User]
      summary: Get patients list
      description: Returns a paginated list of patients.
      operationId: getPatients
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
            example: 1
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            example: 10
        - in: query
          name: selectedText
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Patients retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/deletePatient:
    post:
      tags: [User]
      summary: Delete patient
      description: Soft-deletes a patient and related user records.
      operationId: deletePatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - patientId
              properties:
                patientId:
                  type: string
                  example: PAT-1001
                deletedBy:
                  type: string
                  example: EMP-1001
      responses:
        '200':
          description: Patient deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getPatientProfile:
    get:
      tags: [User]
      summary: Get patient profile by email
      description: Returns patient profile data for the supplied email.
      operationId: getPatientProfile
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Patient profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getPatientId:
    get:
      tags: [User]
      summary: Get patient ID by email
      description: Returns the patient UHID associated with the email.
      operationId: getPatientId
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Patient ID returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Patient id sent successfully
                  patientId:
                    type: string
                    example: PAT-1001
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getAvailableTimeSlots:
    get:
      tags: [User]
      summary: Get available time slots
      description: Returns available appointment slots for a doctor.
      operationId: getAvailableTimeSlots
      parameters:
        - in: query
          name: employeeId
          required: true
          schema:
            type: string
            example: EMP-2001
        - in: query
          name: date
          required: true
          schema:
            type: string
            example: 2026-07-15
      responses:
        '200':
          description: Available slots returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: slots fetched sucessfully
                  slots:
                    type: array
                    items:
                      type: string
                    example: ["10:00", "11:00"]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/updatePatientProfile:
    post:
      tags: [User]
      summary: Update patient profile
      description: Updates patient profile fields by patientId.
      operationId: updatePatientProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdateRequest'
      responses:
        '200':
          description: Patient profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getPatientsBySearch:
    get:
      tags: [User]
      summary: Search patients
      description: Returns patient records matching search text.
      operationId: getPatientsBySearch
      parameters:
        - in: query
          name: searchText
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient search results.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getDoctorsBySearch:
    get:
      tags: [User]
      summary: Search doctors
      description: Returns doctor records matching search text.
      operationId: getDoctorsBySearch
      parameters:
        - in: query
          name: searchText
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Doctor search results.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getPatientById:
    get:
      tags: [User]
      summary: Get patient by ID
      description: Returns a patient record for the supplied patientId.
      operationId: getPatientById
      parameters:
        - in: query
          name: patientId
          required: true
          schema:
            type: string
            example: PAT-1001
      responses:
        '200':
          description: Patient returned successfully.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getDoctorById:
    get:
      tags: [User]
      summary: Get doctor by ID
      description: Returns a doctor record for the supplied doctorId.
      operationId: getDoctorById
      parameters:
        - in: query
          name: doctorId
          required: true
          schema:
            type: string
            example: EMP-2001
      responses:
        '200':
          description: Doctor returned successfully.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/getSingleUser:
    get:
      tags: [User]
      summary: Get a single user by email
      description: Returns combined employee and user profile data.
      operationId: getSingleUser
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User returned successfully.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...